修复列表封面加载失败+1

+++ app/src/main/java/remix/myplayer/db/room/model/PlayList.kt
@@ -26,7 +26,7 @@ data class PlayList(
     @TypeConverter
     fun toStrList(listStr: String?): LinkedHashSet<Long>? {
       val gson = Gson()
-      return gson.fromJson(listStr, object : TypeToken<LinkedHashSet<Int>>() {}.type)
+      return gson.fromJson(listStr, object : TypeToken<LinkedHashSet<Long>>() {}.type)
     }
 
     @TypeConverter
